extract nav and footer to components
