+++ src/components/footer.tsx
@@ -1,0 +1,14 @@
+import React from "react";
+
+const Footer = () => {
+    return (
+        <footer className="footer footer-center p-4 bg-base-300 text-base-content">
+        {/* copyright */}
+        <aside>
+          <p>© 2023 <a href="https://www.mililanihs.org/" className="link">Mililani High School.</a> All Rights Reserved.</p>
+        </aside>
+      </footer>
+    )
+}
+
+export default Footer;
