+++ src/pages/search.tsx
@@ -1,20 +1,8 @@
 import Head from "next/head";
-import Link from "next/link";
-import { useState } from 'react';
-import { useRouter } from "next/router";
+import Navbar from "~/components/navbar";
+import Footer from "~/components/footer";
 
 export default function Home() {
-  const [searchInput, setSearchInput] = useState('');
-  const [selectedProcedure, setSelectedProcedure] = useState('');
-
-  const router = useRouter();
-  const handleButtonClick = async () => {
-    await router.push({
-      pathname: '/search',
-      query: { input: searchInput, procedure: selectedProcedure },
-    });
-  };
-
   return (
     <>
       <Head>
@@ -22,48 +10,7 @@ export default function Home() {
         <meta name="description" content="WIP" />
         <link rel="icon" href="/favicon.ico" />
       </Head>
-      {/**@todo make navbar a component and export */}
-      <div className="navbar bg-base-100">
-        <div className="navbar-start">
-          {/* link to homepage */}
-          <Link href="/">
-            <button className="btn btn-ghost btn-xs sm:btn-sm md:btn-md">
-              <svg height="" viewBox="0 0 1538 531" width="" xmlns="http://www.w3.org/2000/svg"><path d="m648.8 141.1v-74.6h53.9v74.6h44.2v43.4h-44.2v127.4c0 12.2 6 15.6 19.5 15.6 8.2 0 16.9-2.6 24.7-4.3v44.2c-14.7 1.8-27.8 3.5-42.5 3.5-39.9 0-55.6-20-55.6-52v-134.4h-31.2v-43.4m148 261.9c8.2 1.7 14.7 1.7 22.5 1.7 18.7 0 29.1-5.2 35.6-17.8 5.2-10.8 8.7-17.3 8.7-26.4 0-8.2-4.4-17.8-6.9-26.9l-60.4-192.5h54.3l37.3 138.7h.8l38.7-138.7h53.8l-71.2 225.9c-23.8 64.6-43 81.9-87.2 81.9-13 0-19.5-.8-26-1.7zm207.2-349.5h53.8v313.1h-53.8zm138.5 209c.9 46.8 21.7 65 46.9 65 25.1 0 36.8-11.3 50.3-26.9l39 24.3c-20.8 32.1-51.2 46-94.6 46-58.5 0-95.4-45.6-95.4-117.1s36.9-117.1 98-117.1c59.1-.4 92 49.4 92 103.7v22.1zm86.8-36.5c-.9-31.2-17.3-49.4-43.4-49.4-26 0-42.5 18.2-43.4 49.4zm81 140.5v-225.5h49.4v38.7h.9c13-26.9 31.2-42.1 69.4-43.7v56.8c-8.7.3-11.7.3-15.6.3-26.9 0-50.3 30.8-50.3 36.4v137z" fill="#4364a0" fillRule="evenodd" /><path d="m895.3 439.9v-21.8h9.2v21.8h15.2v7.8h-15.2v47c0 3.1 2.1 4.6 7.5 4.6 3 0 5.7-.2 8.1-.6v7.5c-1.6.5-5.3 1-10.7 1-9.9 0-14.1-2.5-14.1-10.9v-48.6h-10.8v-7.8m98.3 51.9c-4.3 7.5-10.4 15.4-25.4 15.4-14.8 0-28.8-10.2-28.8-34.3 0-21.3 10.7-34.3 27-34.3 13.7 0 25.2 8.9 25.2 30v6.1h-43c0 14.5 8.6 24.6 19.6 24.6 8.1 0 14.5-4.3 18.3-11.2zm-11.2-25c0-12.2-5.9-20.3-15.8-20.3-11.1 0-17 8.1-17.8 20.3zm63.3-7.2c-2.5-8-7.8-13.1-15.6-13.1-11.3 0-17.9 9.7-17.9 26.4 0 16.8 6.6 26.4 17.9 26.4 8.3 0 13.1-2.8 18.2-12.1l7.3 3.2c-6.6 11.2-12.4 16.8-25.5 16.8-16.6 0-27.1-13-27.1-34.3s10.5-34.3 27.1-34.3c13.6 0 19.9 7.6 24.1 17.9zm30.7-45.3v34.6h.2c5.2-5.9 12.1-10.2 19.6-10.2 9.8 0 19.1 5 19.1 19.3v48h-9.2v-47.7c0-8.8-4.2-11.7-10.8-11.7-6.8 0-14.7 5.6-18.9 10v49.4h-9.2v-91.7zm64.3 25.7v8.9h.3c5.2-5.9 12.1-10.2 19.6-10.2 9.8 0 19.1 5.1 19.1 19.3v48h-9.2v-47.7c0-8.8-4.2-11.7-10.8-11.7-6.9 0-14.7 5.6-19 10v49.4h-9.1v-66zm78.2 67.3c-14.9 0-27-15.4-27-34.3 0-19 12.1-34.3 27-34.3 15 0 27.1 15.3 27.1 34.3 0 18.9-12.1 34.3-27.1 34.3zm17.9-34.3c0-14.6-8-26.4-17.9-26.4s-17.9 11.8-17.9 26.4 8 26.4 17.9 26.4 17.9-11.8 17.9-26.4zm21.9-58.7h9.2v91.7h-9.2zm49 93c-15 0-27.1-15.3-27.1-34.3s12.1-34.3 27.1-34.3c14.9 0 27 15.3 27 34.3s-12.1 34.3-27 34.3zm17.9-34.3c0-14.6-8-26.4-17.9-26.4-10 0-18 11.8-18 26.4s8 26.4 18 26.4c9.9 0 17.9-11.8 17.9-26.4zm75.6-26.4c-3.8 0-7.8.6-10.1 1.8 1.9 2.2 4.5 7.3 4.5 13.4 0 13.1-10.8 23.2-24.1 23.2-5.3 0-8.1-.8-11.7-1.8-3.6 1-5.2 3.2-5.2 6 0 3.1 2.1 4.9 5.4 4.9 31.3 0 41.4 5.6 41.4 18.5 0 9.7-9.5 17.8-30.2 17.8-18.5 0-28.5-5.1-28.5-16 0-6.6 4.1-9.8 11.3-13.4-5.1-1.6-8.3-5.5-8.3-9.4 0-5.8 3.8-8.2 8.9-11.4-6.2-4.7-9-11.1-9-18.4 0-14 10.9-23.1 25-23.1 7.1 0 12.1 1.5 16.4 4.6 3.9-2.8 9.3-5.4 14.2-5.9zm-31.4 56.7c-14.3 0-18.4 6.3-18.4 11 0 5 3.1 8.3 21.4 8.3 10.9 0 19.8-3.1 19.9-9.6-.1-6.4-6.7-9.7-22.9-9.7zm17.2-41.4c0-8.5-7.3-15.3-16.4-15.3s-16.4 6.8-16.4 15.3c0 8.4 7.3 15.2 16.4 15.2s16.4-6.8 16.4-15.2zm27.3-48.8h9.2v10.9h-9.2zm0 26.9h9.2v66h-9.2zm76 52c-4.4 7.5-10.5 15.4-25.5 15.4-14.7 0-28.7-10.2-28.7-34.3 0-21.4 10.7-34.4 26.9-34.4 13.8 0 25.2 9 25.2 30v6.1h-43c0 14.5 8.7 24.7 19.6 24.7 8.2 0 14.5-4.3 18.3-11.2zm-11.2-25c0-12.2-5.8-20.3-15.7-20.3-11.1 0-17 8.1-17.8 20.3zm27.1 22.3c4.3 6.3 10.6 10.2 18.9 10.2 9.9 0 14.2-5.3 14.2-10.9 0-16.8-39.1-7.9-38.4-32 .4-11.5 9.2-18 21.9-17.8 8.7.1 18.2 4.8 22.6 11.5l-6.3 5.6c-5.1-6.2-10-9.2-17.1-9.2-7.7 0-11.9 3.9-11.9 9.7 0 14.4 38.4 7 38.4 31.7 0 12.8-11.2 19.3-23.2 19.3-10.3 0-20.3-4.5-26.4-12.6z" fill="#939b50" fillRule="evenodd" /><path d="m41.5 224.4c-22.9 0-41.5-18.5-41.5-41.5 0-22.9 18.6-41.5 41.5-41.5 23 0 41.6 18.6 41.6 41.5 0 23-18.6 41.5-41.6 41.5zm70.8 70.7c-23 0-41.6-18.5-41.6-41.5s18.6-41.5 41.6-41.5c22.9 0 41.5 18.5 41.5 41.5s-18.6 41.5-41.5 41.5zm70.7 70.7c-23 0-41.5-18.6-41.5-41.5 0-23 18.5-41.5 41.5-41.5s41.6 18.5 41.6 41.5c0 22.9-18.6 41.5-41.6 41.5zm0 141.3c-23 0-41.5-18.5-41.5-41.5 0-22.9 18.5-41.5 41.5-41.5s41.6 18.6 41.6 41.5c0 23-18.6 41.5-41.6 41.5zm70.7-70.6c-22.9 0-41.5-18.6-41.5-41.6 0-22.9 18.6-41.5 41.5-41.5 23 0 41.6 18.6 41.6 41.5 0 23-18.6 41.6-41.6 41.6zm70.8-70.7c-23 0-41.6-18.6-41.6-41.5 0-23 18.6-41.5 41.6-41.5 22.9 0 41.5 18.5 41.5 41.5 0 22.9-18.6 41.5-41.5 41.5z" fill="#4364a0" /><g fill="#939b50"><path d="m183 224.4c-23 0-41.5-18.5-41.5-41.5 0-22.9 18.5-41.5 41.5-41.5s41.6 18.6 41.6 41.5c0 23-18.6 41.5-41.6 41.5zm70.7-70.6c-22.9 0-41.5-18.6-41.5-41.6 0-22.9 18.6-41.5 41.5-41.5 23 0 41.6 18.6 41.6 41.5 0 23-18.6 41.6-41.6 41.6zm70.8 70.6c-23 0-41.6-18.5-41.6-41.5 0-22.9 18.6-41.5 41.6-41.5 22.9 0 41.5 18.6 41.5 41.5 0 23-18.6 41.5-41.5 41.5zm70.7 70.7c-23 0-41.6-18.5-41.6-41.5s18.6-41.5 41.6-41.5 41.6 18.5 41.6 41.5-18.6 41.5-41.6 41.5zm70.7 70.7c-23 0-41.5-18.6-41.5-41.5 0-23 18.5-41.5 41.5-41.5s41.6 18.5 41.6 41.5c0 22.9-18.6 41.5-41.6 41.5zm-141.4-282.7c-23 0-41.6-18.6-41.6-41.6 0-22.9 18.6-41.5 41.6-41.5 22.9 0 41.5 18.6 41.5 41.5 0 23-18.6 41.6-41.5 41.6z" /><path d="m396.7 388.1c-9.1 0-16.4-7.3-16.4-16.3 0-9.1 7.3-16.4 16.4-16.4s16.4 7.3 16.4 16.4c0 9-7.3 16.3-16.4 16.3zm13.7-16.3c0-7.9-6.1-14.2-13.7-14.2s-13.7 6.3-13.7 14.2c0 7.8 6.1 14.2 13.7 14.2s13.7-6.4 13.7-14.2zm-17.1 9.7h-2.7v-19.5h7.5c4.3 0 6.3 1.9 6.3 5.4s-2.2 5-4.5 5.4l5.4 8.7h-3l-5.3-8.7h-3.7zm3-10.8c2.6 0 5.4.1 5.4-3.3 0-2.8-2.2-3.1-4.2-3.1h-4.2v6.4z" fillRule="evenodd" /></g></svg>
-            </button>
-          </Link>
-          {/* link to map */}
-          <Link href="/map">
-            <button className="btn btn-circle">
-              <svg fill="none" stroke="currentColor" className="h-10 w-10" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
-                <path d="m10 10q0-1.06 0.75-1.81t1.81-0.75 1.81 0.75 0.75 1.81-0.75 1.81-1.81 0.75-1.81-0.75-0.75-1.81zm-2.56 0q0 1.09 0.33 1.79l3.64 7.74q0.16 0.33 0.475 0.52t0.675 0.19 0.675-0.19 0.465-0.52l3.65-7.74q0.33-0.7 0.33-1.79-0-2.12-1.5-3.62t-3.62-1.5-3.62 1.5-1.5 3.62z" strokeLinecap="round" strokeLinejoin="round" strokeWidth="1.2" />
-              </svg>
-            </button>
-          </Link>
-        </div>
-        <div className="navbar-center">
-          {/* search */}
-          <div className="join">
-            {/**@todo: create clear input icon that appears upon user input */}
-            <input className="input input-bordered join-item" placeholder="Search" value={searchInput}
-              onChange={(e) => setSearchInput(e.target.value)} />
-            {/**@todo fix default procedure type */}
-            <select className="select select-bordered join-item" defaultValue={"Procedure Type"}
-              onChange={(e) => setSelectedProcedure(e.target.value)}>
-              <option disabled>Procedure Type</option>
-              <option>Family Care</option>
-              <option>Radiology</option>
-              <option>...</option>
-            </select>
-            <button className="btn btn-success btn-outline btn-circle join-item" onClick={handleButtonClick}>
-              <svg xmlns="http://www.w3.org/2000/svg" className="h-5 w-5" fill="none" viewBox="0 0 24 24" stroke="currentColor"><path strokeLinecap="round" strokeLinejoin="round" strokeWidth="2" d="M21 21l-6-6m2-5a7 7 0 11-14 0 7 7 0 0114 0z" /></svg>
-            </button>
-          </div>
-        </div>
-        <div className="navbar-end">
-          {/* link */}
-          <a className="btn btn-ghost normal-case text-xl" href="https://github.com/jhung-mililani/MHSHAKK">MHS-HAKK</a>
-        </div>
-      </div>
+      <Navbar></Navbar>
       <main className="flex min-h-screen flex-col items-center justify-center bg-gradient-to-b from-[#2e026d] to-[#15162c]">
         <div className="container flex flex-col items-center justify-center gap-12 px-4 py-16 ">
           <h1 className="text-5xl font-extrabold tracking-tight text-white sm:text-[5rem]">
@@ -71,6 +18,7 @@ export default function Home() {
           </h1>
         </div>
       </main>
+      <Footer></Footer>
     </>
   );
 }
